+++ src/main.jsx
@@ -7,19 +7,11 @@ import NotFound from './NotFound.jsx';
 import PantanoDetails from './PantanoDetails.jsx';
 import Perfil from './Perfil.jsx';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { UserContext } from './context/UserContext';
+import { UserProvider } from './context/UserContext';
 import Login from './Login.jsx';
 import Registro from './Registro.jsx';
 
 //console.log(import.meta.env.VITE_PROJECT_ID)
-
-// Usuario simulado
-const usuarioSimulado = {
-  id: "Ndw9ViAft6gPct2roo5A ",
-  nombre: "Javier Ruiz",
-  email: "javier@gmail.com",
-  telefono: "610234567"
-};
 
 const router = createBrowserRouter([
   { path: "/", element: <Home /> },
@@ -33,8 +25,8 @@ const router = createBrowserRouter([
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
-    <UserContext.Provider value={usuarioSimulado}>
+    <UserProvider>
       <RouterProvider router={router} />
-    </UserContext.Provider>
+    </UserProvider>
   </StrictMode>
 );
